Removed unused classes Changed value from Double to BigDecimal Changed Ids from Integer To Long Adjust code to changes

+++ src/main/java/com/expensetracker/controller/ExpenseController.java
@@ -33,7 +33,7 @@ public class ExpenseController {
     }
 
     @DeleteMapping("/expenses/{id}")
-    ResponseEntity<?> deleteExpense(@PathVariable Integer id) {
+    ResponseEntity<?> deleteExpense(@PathVariable Long id) {
         expenseService.deleteExpenseById(id);
 
         return ResponseEntity.ok().build();
@@ -46,7 +46,7 @@ public class ExpenseController {
     }
 
     @PostMapping("/expenses/{id}")
-    ResponseEntity<Expense> duplicateExpense(@PathVariable(name = "id") Integer expenseId) throws URISyntaxException {
+    ResponseEntity<Expense> duplicateExpense(@PathVariable(name = "id") Long expenseId) throws URISyntaxException {
         Expense result = expenseService.duplicateExpense(expenseId);
         return ResponseEntity.created(new URI("/api/expenses" + result.getId())).body(result);
     }
